requirements.txt
streamlit
scikit-learn==1.3.2
plotly
openai
python-dotenv